requirements.txt
# Object Detection Course

cvzone==1.5.6
ultralytics==8.0.121
hydra-core>=1.2.0
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
filterpy==1.4.5
scikit-image==0.19.3
PyJWT~=2.7.0
Flask~=2.3.2
Werkzeug~=2.3.6
alembic~=1.11.1
SQLAlchemy~=2.0.16
opencv-python~=4.7.0.72
imutils~=0.5.4